# openapi.yaml
openapi: 3.1.0
info:
  title: Historiador Doc API
  description: REST API for Historiador Doc — self-hosted documentation with a built-in MCP server.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
  version: 0.1.0
paths:
...
  /admin/users:
    get:
      tags:
      - admin
      operationId: list_users
      responses:
        '200':
          description: user list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: unauthenticated
        '403':
          description: caller is not admin
      security:
      - bearer: []
  /admin/users/invite:
    post:
      tags:
      - admin
      operationId: invite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
        required: true
      responses:
        '200':
          description: invited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          description: validation error
        '401':
          description: unauthenticated
        '403':
          description: caller is not admin
        '409':
          description: email already in use
      security:
      - bearer: []
  /admin/users/{id}/deactivate:
    patch:
      tags:
      - admin
      operationId: deactivate_user
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: user deactivated
        '401':
          description: unauthenticated
        '403':
          description: caller is not admin
        '404':
          description: user not found
      security:
      - bearer: []
...
  /auth/activate:
    post:
      tags:
      - auth
      operationId: activate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateRequest'
        required: true
      responses:
        '204':
          description: activated
        '400':
          description: validation error
        '401':
          description: invite token invalid or expired
...
  /editor/draft:
    post:
      tags:
      - editor
      operationId: draft
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftRequest'
        required: true
      responses:
        '200':
          description: 'SSE stream of generated markdown; event types: `delta` (data: {"text": "..."}), `error` (data: {"message": "..."}), `done` (data: {"length": N}). Content-Type: text/event-stream.'
          content:
            text/event-stream: {}
        '400':
          description: validation error
        '401':
          description: unauthorized
        '403':
          description: forbidden
      security:
      - bearer: []
  /editor/iterate:
    post:
      tags:
      - editor
      operationId: iterate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IterateRequest'
        required: true
      responses:
        '200':
          description: SSE stream of refined markdown; see /editor/draft for event shape.
          content:
            text/event-stream: {}
        '400':
          description: validation error
        '401':
          description: unauthorized
        '403':
          description: forbidden
      security:
      - bearer: []
...
  /pages/{id}/editor-conversation:
    get:
      tags:
      - editor
      operationId: get_conversation
      parameters:
      - name: id
        in: path
        description: page id
        required: true
        schema:
          type: string
          format: uuid
      - name: language
        in: query
        description: |-
          BCP 47 language tag identifying which language the conversation
          belongs to, since authors can hold parallel threads per language.
        required: true
        schema:
          type: string
        example: en
      responses:
        '200':
          description: Persisted conversation or an empty transcript if none saved yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: page not found
      security:
      - bearer: []
    put:
      tags:
      - editor
      operationId: put_conversation
      parameters:
      - name: id
        in: path
        description: page id
        required: true
        schema:
          type: string
          format: uuid
      - name: language
        in: query
        description: |-
          BCP 47 language tag identifying which language the conversation
          belongs to, since authors can hold parallel threads per language.
        required: true
        schema:
          type: string
        example: en
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveConversationRequest'
        required: true
      responses:
        '200':
          description: Updated conversation echo (client uses `updated_at` for the banner).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: validation error
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: page not found
      security:
      - bearer: []
...
components:
  schemas:
    ActivateRequest:
      type: object
      required:
      - invite_token
      - password
      properties:
        invite_token:
          type: string
        password:
          type: string
...
    ConversationMessageDto:
      type: object
      required:
      - role
      - content
      - ts
      properties:
        content:
          type: string
        role:
          type: string
          description: '"user" or "assistant".'
        ts:
          type: integer
          format: int64
          description: Client-side millisecond Unix timestamp.
    ConversationResponse:
      type: object
      required:
      - page_id
      - language
      - messages
      - updated_at
      properties:
        language:
          type: string
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessageDto'
        page_id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
...
    DraftRequest:
      type: object
      required:
      - brief
      properties:
        brief:
          type: string
        language:
          type:
          - string
          - 'null'
...
    InviteRequest:
      type: object
      required:
      - email
      - role
      properties:
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
    InviteResponse:
      type: object
      required:
      - user_id
      - activation_url
      - expires_at
      properties:
        activation_url:
          type: string
        expires_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
    IterateRequest:
      type: object
      required:
      - current_draft
      - instruction
      properties:
        current_draft:
          type: string
        instruction:
          type: string
...
    Role:
      type: string
      enum:
      - admin
      - author
      - viewer
    SaveConversationRequest:
      type: object
      required:
      - messages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessageDto'
...
    UserResponse:
      type: object
      required:
      - id
      - email
      - role
      - active
      - pending
      properties:
        active:
          type: boolean
        email:
          type: string
        id:
          type: string
          format: uuid
        pending:
          type: boolean
          description: '`true` if the user has not yet activated their account.'
        role:
          $ref: '#/components/schemas/Role'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...